requirements.txt
alembic==1.12.1
asttokens==2.4.1
backcall==0.2.0
bleach==6.1.0
blinker==1.6.3
bokeh==3.3.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==2.2.1
colorcet==3.0.1
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
databricks-cli==0.18.0
debugpy==1.8.0
decorator==5.1.1
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.0
Flask==2.3.3
fonttools==4.43.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
gunicorn==21.2.0
pandas==2.1.2
polars==0.19.12
scikit-learn==1.3.2
panel==1.3.0
hvplot==0.9.0
holoviews==1.18.0
huggingface-hub==0.17.3
pandas==2.1.2
polars==0.19.12
scikit-learn==1.3.2
panel==1.3.0
hvplot==0.9.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.5.0
jupyter_core==5.4.0
kiwisolver==1.4.5
lightning-utilities==0.9.0
linkify-it-py==2.0.2
Mako==1.2.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mlflow==2.7.1
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.2
pandas==2.1.2
polars==0.19.12
scikit-learn==1.3.2
panel==1.3.0
param==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
pandas==2.1.2
polars==0.19.12
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyct==0.5.0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyviz_comms==3.0.0
PyYAML==6.0.1
pyzmq==25.1.1
querystring-parser==1.2.4
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
pandas==2.1.2
polars==0.19.12
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.22
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
torchmetrics==1.2.0
torchvision==0.16.0
tornado==6.3.3
pandas==2.1.2
polars==0.19.12
scikit-learn==1.3.2
panel==1.3.0
hvplot==0.9.0
holoviews==1.18.0
transformers==4.34.1
tqdm==4.66.1
traitlets==5.12.0
pandas==2.1.2
polars==0.19.12
scikit-learn==1.3.2
panel==1.3.0
hvplot==0.9.0
holoviews==1.18.0
transformers==4.34.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.7
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
xyzservices==2023.10.1
zipp==3.17.0